afsutil.package: save generated rpms when building with mock
Save the generate rpms to the package/rpmbuild/RPMS/<arch>
directory by default when building with mock. This destination
path can be overridded with the dstdir option.

Add the 'sources' --build target to build just the spec file
and the tarballs.

Pass the --verbose option to mock for extra messages.