package: build kmods in order
Simplify how we skip already built kmods and be sure to build
them in the same order as listed/found.  Change repoid option
to --khrepo (short for kernel headers repo) and default to the
value 'kernel-headers'.  Fix --list-kversions to print just the
versions to stdout as a simple list.  Dont include errors from
repoquery as versions.