Error out if unable to do build --clean
The git clean is now explicit with the build --clean option, so error
out if we can not run the git clean because we are not in the git
directory or the afsutil.clean git option is not enabled, instead
of printing a info message and continuing.

Also, do not fall back to make clean, since that is not sufficent when
changing branches, just always run git clean when --clean is specified.