package: remove --khrepo option, add --chroot option
Remove the awkward --khrepo which blocked us for just
querying all of the yum repos for kernel-devel packages.

Add the --chroot option and make it default to "default"
which is the same as the mock --root default. Change --mock
to be just a boolean switch to indicate if we are in mock
mode or not.