package: check for dashes in package version and release
The rpm package version and release values may not contain dash (-)
characters, since rpm reserves the dash character as a delimiter.
Try to catch invalid version names early since it is a common error,
instead of waiting for the rpmbuild to fail.

Set the package version to "99.99.88" and the release the abbreviated
sha1 when it appears we are packaging a checked out master branch. In
this case we are building rpms for testing.

R remove any stray dashes from the package version when the version
string looks like a custom format