configure: fallback to python2 and pip2
Starting with RHEL/CentOS 8, python is called python2.

When python is not present, try to find and use python2
in the make process.  The setup.py will automatically
configure the afutil shebang line for us during the setup.

Set the configured python bin to be used when building the rpm,
instead of the default "python" in bdist_rpm. This will set
python2 in the shebang of the installed afsutil script, and
cause python2 to be a requirement for the rpm on RHEL/CentOS 8.